Removed hard coded CompositeActivity states and transitions
Replaced autorun logic with tests on initialState and isFinishing, so
that custom composite activity state machines will work.
If more than one transition is available on run() then it will not
auto-execute. runNext() will override a finishing() transition with a
non finishing, but will give up if there is more than one of each.